requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-taggit==2.0.0
environs==9.4.0
marshmallow==3.14.1
Pillow==9.0.0
psycopg2==2.9.3
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
